채팅 모듈 테스트 파일 구현

+++ src/providers/database/mongoose/mongoose.seed.ts
@@ -5,12 +5,15 @@ import { ChatModel } from './chat.schema';
 import CHAT_ROOMS from './mock/chatRoom.mock';
 import { NotificationModel } from './notification.schema';
 import NOTIFICATIONS from './mock/notification.mock';
+import CHATS from './mock/chat.mock';
 
 async function connectDB() {
   await mongoose.connect(process.env.MONGO_URI);
 }
+let isSeeded = false;
+export async function mongooseSeed() {
+  if (isSeeded) return;
 
-export async function seed() {
   await connectDB();
 
   await ChatModel.deleteMany();
@@ -18,13 +21,15 @@ export async function seed() {
   await NotificationModel.deleteMany();
 
   await ChatRoomModel.insertMany(CHAT_ROOMS);
+  await ChatModel.insertMany(CHATS);
   await NotificationModel.insertMany(NOTIFICATIONS);
 
   console.log('🌱 Mongoose Seeding completed!');
+  isSeeded = true;
   mongoose.connection.close();
 }
 
-seed().catch((e) => {
+mongooseSeed().catch((e) => {
   console.error(e);
   mongoose.connection.close();
 });
